temporary fix for githook to run
